requirements.txt
streamlit==1.42.0
langchain-groq==0.2.4
langgraph==0.2.68
spacy==3.7.6
sentence-transformers==3.4.1
python-dotenv==1.0.1
ddgs
langgraph-checkpoint==2.0.11
spacy
requests
hypothesis
pytest
sentence-transformers
rouge-score
textblob
pydantic
